Moved apollo methods in extension: InterventionViewController+Apollo. Added articles saving methods
